backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
google-generativeai
psycopg2-binary
sqlalchemy
pgvector
langchain-community 